Updated deliveryLocation function.
It was observed that despite maximumAge being set to 0 the initial result from watchPosition sometimes was a previous high accuracy location causing the location to not update if the accuracy was not sufficient. To mitigate this the first 2 results are now discarded assuming that by the third attempt the device will have a sufficient fix on its location. Additionally, to cut down on wait time, the first result after the second that has an accuracy of less than 100 meters is returned.